final changes for doing decoding and alignment of native recordings
Former-commit-id: bac45fb63c4c36b688510dd4a17c4af89fd2dab2 [formerly 69a8163d29ea0894b35880aab8c46b08f6b0de86]
Former-commit-id: 6f5d893b49cf7449222a0973b7af2f26c3a8ae0c [formerly 4cf33646c76e3f1b347b72635e5c6c7c45ccd75b]
Former-commit-id: a032b6cb7146bd63ced5e2f0b703e7cc21b00797

+++ src/mitll/langtest/shared/AudioAnswer.java
@@ -25,7 +25,7 @@ public class AudioAnswer implements IsSerializable {
   private double score = -1;
   private boolean correct = false;
   private boolean saidAnswer = false;
-  private int durationInMillis;
+  private long durationInMillis;
   private long resultID;
   private AudioAttribute audioAttribute;
   private PretestScore pretestScore;
@@ -56,7 +56,7 @@ public class AudioAnswer implements IsSerializable {
    * @param reqid
    * @param duration
    */
-  public AudioAnswer(String path, Validity validity, int reqid, int duration) {
+  public AudioAnswer(String path, Validity validity, int reqid, long duration) {
     this.path = path;
     this.validity = validity;
     this.reqid = reqid;
@@ -79,7 +79,6 @@ public class AudioAnswer implements IsSerializable {
 
   /**
    * @see mitll.langtest.client.recorder.RecordButtonPanel#receivedAudioAnswer(AudioAnswer, com.google.gwt.user.client.ui.Panel)
-   * @see mitll.langtest.server.autocrt.AutoCRT#getAutoCRTDecodeOutput(String, int, java.io.File, AudioAnswer)
    * @return
    */
   public boolean isCorrect() { return correct; }
@@ -130,7 +129,7 @@ public class AudioAnswer implements IsSerializable {
     return decodeOutput;
   }
 
-  public int getDurationInMillis() {
+  public long getDurationInMillis() {
     return durationInMillis;
   }
